Extract MakeTimePoint into time_testing.h
... for reuse elsewhere
